// frontend/lib/graphql/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  ISO8601Date: { input: any; output: any; }
};
// ...
export type DishSourceRelation = {
  __typename?: 'DishSourceRelation';
  dishId: Scalars['Int']['output'];
  dishSourceId: Scalars['Int']['output'];
  recipeBookPage?: Maybe<Scalars['Int']['output']>;
  recipeSourceMemo?: Maybe<Scalars['String']['output']>;
  recipeWebsiteUrl?: Maybe<Scalars['String']['output']>;
  sourceName?: Maybe<Scalars['String']['output']>;
  type: Scalars['Int']['output'];
};